backend/requirements-dev.txt
flask==3.0.0
black==23.11.0
flake8==6.1.0
pyright==1.1.336
pydantic==2.5
pytest==7.3.1
psycopg2-binary==2.9.9
wheel
flask-restful==0.3.10
